Author Post permission Fixed

+++ app/Http/Controllers/Author/PostController.php
@@ -105,6 +105,12 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+
+        if($post->user_id != Auth::id()){
+            Toastr::error('You are not authorized to access this page', 'Error');
+            return redirect()->back();
+        }
+
         return view('author.post.show', compact(
             'post'
         ));
@@ -118,6 +124,11 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error('You are not authorized to access this page', 'Error');
+            return redirect()->back();
+        }
+
         $categories = Category::all();
         $tags       = Tag::all();
         return view('author.post.edit', compact('post', 'categories', 'tags'));
@@ -132,6 +143,11 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error('You are not authorized to access this page', 'Error');
+            return redirect()->back();
+        }
+
         $request->validate([
             'title'     => 'required',
             'image'     => 'image',
@@ -163,7 +179,7 @@ class PostController extends Controller
             $imageName = $post->image;
         }
 
-        $post->user_id = Auth::id();
+        $post->user_id = $post->user_id;
         $post->title   = $request->title;
         $post->slug    = $slug;
         $post->image   = $imageName;
@@ -191,6 +207,11 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        if($post->user_id != Auth::id()){
+            Toastr::error('You are not authorized to access this page', 'Error');
+            return redirect()->back();
+        }
+        
         if(Storage::disk('public')->exists('post/'.$post->image)){
             Storage::disk('public')->delete('post/'.$post->image);
         }
